TJFE-8 [rename] HomeScreen 패키지 경로 수정
